make `check_instance()` default to `True` for unhandled generic types
If a type annotation is a generic, but not a dict, list, tuple, or set, the `check_instance()` function ends up calling `isinstance(obj, type_)`. This causes an error for other generic types. For instance, when using `numpy.typing`, the type NDArray[np.float_])` results in the following exception:

`TypeError: isinstance() argument 2 cannot be a parameterized generic`.

This change at least avoids that error, but it defaults to always returning `True`, which does not check any of the element types. So the type checking is incomplete. However, unless Hamilton wants to implement a full type checker, that's probably unavoidable.

For full runtime type checking, it probably makes more sense to use something like typeguard or jaxtyping.